Change SocketAddr types structure.
The problem with using an enum `SocketAddr` to store the different types of `struct socketaddr` is that, because Rust uses tagged unions, you cannot convert the SocketAddr to the `struct socketaddr`.
For this reason, handling the addresses would require a lot of switches.
Using generic functions, instead, might be more verbose, but ensures that the parameters and returns type are of the concrete socket address type, which is transparent and therefore castable directly as a socketaddr.
However, this change requires some further review, but I pretty much like the ease of use.

Signed-off-by: Michele Dalle Rive <dallem@usi.ch>