Add Socket receive and send functions.
The receive and send messages have the problem of being protocol-dependent. While the other methods behave, more or less, the same, receive and send behave drastically differently depending on the protocol: for example, in TCP, receive just retrieves the bytes and modifies the buffer; UDP, instead, must also communicate the address of the sender, since it is not connection-oriented. For this reason, the `msghdr` mechanism is used, which stinks a bit. Therefore, there will be probably socket wrappers for tcp and udp that return the protocol-specific required fields.

Signed-off-by: Michele Dalle Rive <dallem@usi.ch>