Remove `msghdr` Rust wrapper.
Having a Rust wrapper for msghdr could be useful in order to give a certain level of freedom on the message that is sent. The most important field, msg_name, which contains a pointer to a sockaddr_* struct, gives the most problems.
In order to have a "movable-around" structure, the sockaddr object must be stored inside the Message object itself, otherwise the reference could expire while the object is still alive.
This implies having a self-referencing structure: the msghdr field must point to the sockaddr field. This can be solved by pinning the structure.
However, the implementation and usage complexity of such a structure gives no advantage, since the only relevant fields (as far as I know) are the msg_name, msg_namelen and msg_flags fields, which can be set manually when calling the set and get methods.
This way, the reference is obviously valid and there is no pinning problem.
KISS

Signed-off-by: Michele Dalle Rive <dallem@usi.ch>