rust: net: add net module files and shared enums.
The enums added with this patch are base enums shared across the whole network subsystem. For this reason, they are placed in the `net.rs` module file.

The enum IpProtocol, however, is placed in an individual `ip.rs` submodule, allowing to place together all the ip-related structures, such as wrappers for `iphdr`, `ip_auth_hdr`, etc.

Signed-off-by: Michele Dalle Rive <dallerivemichele@gmail.com>