rust: net: add socket wrapper.
These bindings allow to use the kernel socket API through a Rust `Socket` abstraction.

The Socket struct tries to keep the same function signatures of the Rust standard library.
At the same time, functions are added or modified in order to provide as much as possible of the C kernel functionalities.

Most of the internals of the C socket is not accessible by Rust, because those structs are still to be wrapped. However, the use of socket happens mainly through the functions, thus most of the functionality should be available in Rust as well.

Specifically, the usage of `msghdr` is mostly abstracted away in the Rust interface, because using it would mean having to deal both in the kernel and in modules with Pinned instances (msghdr is self-referencing), which would be a struggle that provides no particular advantage.
However, some useful functionalities of `msghdr`, like `cmsghdr`s, are missing and should be implemented in the future to provide a complete API.

Signed-off-by: Michele Dalle Rive <dallerivemichele@gmail.com>