rust: net: implement socket options API.
These changes introduce wrappers and functions to handle socket options in Rust, with compilation-time advantages compared to the C API:
- Type safety: A specific option accepts only a value of the correct type.
- Read/write safety: A read-only option cannot be set.
- Coherence safety: An option of IP level cannot be set by specifying another level.

The downside of using options in the kernel is the lack of a way to get the value of an option. For this reason, in Rust, kernel options can only be set but not retrieved.

Everything that can be done by socket options can be done through functions or by accessing directly the specific fields. However, since the Rust-wrapped structures are few, it can be useful to have options in order to still be able to modify the behaviour of the socket.

As specified in the documentation of `opts.rs`, options could be removed in the future, when the Rust API will be developed enough.

Signed-off-by: Michele Dalle Rive <dallerivemichele@gmail.com>